requirements.txt
Flask
numpy
opencv-python
Pillow
openai
python-dotenv
